chore: move attributes queries to folder
