+++ src/pages/SettingsPage/Attributes.jsx
@@ -4,8 +4,8 @@ import { DataTable } from 'primereact/datatable'
 import { Column } from 'primereact/column'
 import { TablePanel, Button, Section, Toolbar, InputText } from '@ynput/ayon-react-components'
 import AttributeEditor from '../../containers/attributes/attributeEditor'
-import { useGetAttributesQuery } from '/src/services/getAttributes'
-import { useUpdateAttributesMutation } from '/src/services/updateAttributes'
+import { useGetAttributesQuery } from '/src/services/attributes/getAttributes'
+import { useUpdateAttributesMutation } from '/src/services/attributes/updateAttributes'
 import useSearchFilter from '/src/hooks/useSearchFilter'
 import { ConfirmDialog, confirmDialog } from 'primereact/confirmdialog'
 import axios from 'axios'
